+++ company-service/src/main/java/com/oneforlogis/company/application/CompanyService.java
@@ -5,10 +5,10 @@ import com.oneforlogis.common.exception.ErrorCode;
 import com.oneforlogis.company.domain.model.Company;
 import com.oneforlogis.company.domain.model.CompanyType;
 import com.oneforlogis.company.domain.repository.CompanyRepository;
-import com.oneforlogis.company.presentation.dto.request.CompanyCreateRequest;
-import com.oneforlogis.company.presentation.dto.request.CompanyUpdateRequest;
-import com.oneforlogis.company.presentation.dto.response.CompanyCreateResponse;
-import com.oneforlogis.company.presentation.dto.response.CompanyUpdateResponse;
+import com.oneforlogis.company.application.dto.request.CompanyCreateRequest;
+import com.oneforlogis.company.application.dto.request.CompanyUpdateRequest;
+import com.oneforlogis.company.application.dto.response.CompanyCreateResponse;
+import com.oneforlogis.company.application.dto.response.CompanyUpdateResponse;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
