--- company-service/src/main/java/com/oneforlogis/company/application/dto/request/request.java
@@ -1,2 +1,0 @@
-package com.oneforlogis.company.application.dto.request;
-// request dto
