+++ company-service/src/main/java/com/oneforlogis/company/presentation/controller/CompanyController.java
@@ -3,10 +3,10 @@ package com.oneforlogis.company.presentation.controller;
 import com.oneforlogis.common.api.ApiResponse;
 import com.oneforlogis.common.security.UserPrincipal;
 import com.oneforlogis.company.application.CompanyService;
-import com.oneforlogis.company.presentation.dto.request.CompanyCreateRequest;
-import com.oneforlogis.company.presentation.dto.request.CompanyUpdateRequest;
-import com.oneforlogis.company.presentation.dto.response.CompanyCreateResponse;
-import com.oneforlogis.company.presentation.dto.response.CompanyUpdateResponse;
+import com.oneforlogis.company.application.dto.request.CompanyCreateRequest;
+import com.oneforlogis.company.application.dto.request.CompanyUpdateRequest;
+import com.oneforlogis.company.application.dto.response.CompanyCreateResponse;
+import com.oneforlogis.company.application.dto.response.CompanyUpdateResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
